third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-blocklisted-characteristic-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  return getHIDDevice({filters: [{services: ['device_information']}]})
    .then(({device}) => device.gatt.getPrimaryService('device_information'))
    .then(service => {
      return assert_promise_rejects_with_message(
        service.getCharacteristics('serial_number_string'),
        new DOMException('getCharacteristic(s) called with blocklisted UUID. ' +
                         'https://goo.gl/4NeimX',
                         'SecurityError'),
        'Serial Number String characteristic is blocklisted.');
    });
}, 'Serial Number String characteristic is blocklisted. Should reject with ' +
   'SecurityError.');

</script>
